reorganizing files and changing names for clearer code
