+++ js/api/api-authenticate/signup/login-account.mjs
@@ -1,4 +1,4 @@
-import { authenticateProfile } from "../api-authenticate/authenticate-profile.mjs";
+import { authenticateProfile } from "../authenticate-profile.mjs";
 
 export function loginAccount(event) {
    const formElements = event.target;
