+++ js/display/post/display-profile-data.mjs
@@ -1,5 +1,5 @@
-import { changeAvatarData } from "../api/change-api-data/change-avatar-data.mjs";
-import { loadItem } from "../storage/local-storage.mjs";
+import { changeAvatarData } from "../../api/change-api-data/change-avatar-data.mjs";
+import { loadItem } from "../../storage/local-storage.mjs";
 
 export function displayProfileData() {
    const profile = loadItem("profile");
