+++ js/display/signup/signup-feedback/display-login-feedback.mjs
@@ -1,7 +1,7 @@
-import { redirectIfToken } from "../api/api-authenticate/redirect-active-user.mjs";
-import { createParseDoc } from "../html-data/createParseDoc.mjs";
-import { saveItem } from "../storage/local-storage.mjs";
-import { saveSessionItem } from "../storage/session-storage.mjs";
+import { redirectIfToken } from "../../../api/api-authenticate/redirect-active-user.mjs";
+import { createParseDoc } from "../../../html-data/createParseDoc.mjs";
+import { saveItem } from "../../../storage/local-storage.mjs";
+import { saveSessionItem } from "../../../storage/session-storage.mjs";
 
 export function displayLoginFeedback(response, result) {
    // Save login token & profile
