+++ js/handlers/posts/create-post-handler.mjs
@@ -1,5 +1,5 @@
 import { createRequest } from "../../api/api-requests/request-type/create-request.mjs";
-import { validateInputs } from "../../form-validation/validatedInput.mjs";
+import { validateInputs } from "../../form/validatedInput.mjs";
 
 export function createPostHandler() {
    const createPostForm = document.querySelector("#create-post-form");
